+++ code/sonalyze/src/prjobs.rs
@@ -393,7 +393,7 @@ fn format_duration_sec(JobSummary { aggregate: a, .. }: LogDatum, _: LogCtx) -> 
 fn format_cputime(JobSummary { aggregate: a, .. }: LogDatum, _: LogCtx) -> String {
     // See below for a description of the computation.
     let duration = 60 * (a.minutes + 60 * (a.hours + (a.days * 24)));
-    let mut cputime_sec = (a.cpu_avg * duration as f64 / 100.0).round() as i64;
+    let mut cputime_sec = (a.cpu_avg * duration as f32 / 100.0).round() as i64;
     if cputime_sec % 60 >= 30 {
         cputime_sec += 30;
     }
@@ -409,14 +409,14 @@ fn format_cputime_sec(JobSummary { aggregate: a, .. }: LogDatum, _: LogCtx) -> S
     // (whose units is second) to get total core-seconds for the job.  Finally scale by 100 because
     // the cpu_avg numbers are expressed in integer percentage point.
     let duration = 60 * (a.minutes + 60 * (a.hours + (a.days * 24)));
-    let cputime = (a.cpu_avg * duration as f64 / 100.0).round() as i64;
+    let cputime = (a.cpu_avg * duration as f32 / 100.0).round() as i64;
     format!("{}", cputime)
 }
 
 fn format_gputime(JobSummary { aggregate: a, .. }: LogDatum, _: LogCtx) -> String {
     // See below for a description of the computation.
     let duration = 60 * (a.minutes + 60 * (a.hours + (a.days * 24)));
-    let mut gputime_sec = (a.gpu_avg * duration as f64 / 100.0).round() as i64;
+    let mut gputime_sec = (a.gpu_avg * duration as f32 / 100.0).round() as i64;
     if gputime_sec % 60 >= 30 {
         gputime_sec += 30;
     }
@@ -432,7 +432,7 @@ fn format_gputime_sec(JobSummary { aggregate: a, .. }: LogDatum, _: LogCtx) -> S
     // (whose units is second) to get total card-seconds for the job.  Finally scale by 100 because
     // the gpu_avg numbers are expressed in integer percentage point.
     let duration = 60 * (a.minutes + 60 * (a.hours + (a.days * 24)));
-    let gputime = (a.gpu_avg * duration as f64 / 100.0).round() as i64;
+    let gputime = (a.gpu_avg * duration as f32 / 100.0).round() as i64;
     format!("{}", gputime)
 }
